chore: rename and move receive screen file
